+++ src/app/item-popup/ItemStat.tsx
@@ -366,7 +366,9 @@ function getNonReusableModSockets(item: DimItem) {
       !socketContainsIntrinsicPlug(s) &&
       !s.plugged.plugDef.plug.plugCategoryIdentifier.includes('masterwork') &&
       (s.plugged.plugDef.itemCategoryHashes?.some((h) => modItemCategoryHashes.has(h)) ||
-        statfulOrnaments.includes(s.plugged.plugDef.hash)),
+        statfulOrnaments.includes(s.plugged.plugDef.hash) ||
+        // Tuning mods don't have the ArmorMods item category hash
+        s.plugged.plugDef.plug.plugCategoryIdentifier.includes('tuning.mods')),
   );
 }
 /**