+++ src/app/item-popup/PlugTooltip.tsx
@@ -25,7 +25,7 @@ import {
 import clsx from 'clsx';
 import enhancedIntrinsics from 'data/d2/crafting-enhanced-intrinsics';
 import { TraitHashes } from 'data/d2/generated-enums';
-import { useCallback } from 'react';
+import React, { useCallback } from 'react';
 import { DimItem, DimPlug, PluggableInventoryItemDefinition } from '../inventory/item-types';
 import Objective from '../progress/Objective';
 import styles from './PlugTooltip.m.scss';
@@ -147,13 +147,17 @@ function PlugTooltip({
   const bungieDescription =
     plugDescriptions.perks.length > 0 &&
     plugDescriptions.perks.map((perkDesc) => (
-      <div key={perkDesc.perkHash}>
-        {perkDesc.name && <div className={styles.perkName}>{perkDesc.name}</div>}
-        {perkDesc.description && <RichDestinyText text={perkDesc.description} />}
+      <React.Fragment key={perkDesc.perkHash}>
+        {perkDesc.name && !perkDesc.name.match(/^\+.*▲/m) && (
+          <div className={styles.perkName}>{perkDesc.name}</div>
+        )}
+        {perkDesc.description && perkDesc.description !== perkDesc.name && (
+          <RichDestinyText text={perkDesc.description} />
+        )}
         {!hideRequirements && perkDesc.requirement && (
           <RichDestinyText text={perkDesc.requirement} className={styles.requirement} />
         )}
-      </div>
+      </React.Fragment>
     ));
   const clarityDescriptionSection = plugDescriptions.communityInsight && (
     <Tooltip.Section className={styles.communityInsightSection}>